+++ src/main/java/com/yimei/hs/mapper/CangFukuanMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangFukuan;
 public interface CangFukuanMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(CangFukuan record);
+    long insert(CangFukuan record);
 
     int insertSelective(CangFukuan record);
